+++ server/knowledge_base/kb_service/base.py
@@ -19,16 +19,16 @@ from server.db.repository.knowledge_file_repository import (
 )
 
 from configs import (kbs_config, VECTOR_SEARCH_TOP_K, SCORE_THRESHOLD,
-                     EMBEDDING_MODEL, KB_INFO, logger)
+                     EMBEDDING_MODEL, KB_INFO)
 from server.knowledge_base.utils import (
     get_kb_path, get_doc_path, KnowledgeFile,
     list_kbs_from_folder, list_files_from_folder,
 )
-from server.utils import (get_model_worker_config, BaseResponse,
-                         list_embed_models, list_online_embed_models)
-from server.model_workers.base import ApiEmbeddingsParams, ApiModelWorker
 
-from typing import List, Union, Dict, Optional, Tuple
+from typing import List, Union, Dict, Optional
+
+from server.embeddings_api import embed_texts
+from server.embeddings_api import embed_documents
 
 
 class SupportedVSType:
@@ -385,52 +385,3 @@ def score_threshold_process(score_threshold, k, docs):
             if cmp(similarity, score_threshold)
         ]
     return docs[:k]
-
-
-def embed_texts(
-    texts: List[str],
-    embed_model: str = EMBEDDING_MODEL,
-    to_query: bool = False,
-) -> BaseResponse:
-    '''
-    对文本进行向量化。返回数据格式：BaseResponse(data=List[List[float]])
-    '''
-    try:
-        if embed_model in list_embed_models(): # 使用本地Embeddings模型
-            from server.utils import load_local_embeddings
-
-            embeddings = load_local_embeddings(model=embed_model)
-            return BaseResponse(data=embeddings.embed_documents(texts))
-        
-        if embed_model in list_online_embed_models(): # 使用在线API
-            config = get_model_worker_config(embed_model)
-            worker_class = config.get("worker_class")
-            worker = worker_class()
-            if worker_class.can_embedding():
-                params = ApiEmbeddingsParams(texts=texts, to_query=to_query)
-                resp = worker.do_embeddings(params)
-                return BaseResponse(**resp)
-
-        return BaseResponse(code=500, msg=f"指定的模型 {embed_model} 不支持 Embeddings 功能。")
-    except Exception as e:
-        logger.error(e)
-        return BaseResponse(code=500, msg=f"文本向量化过程中出现错误：{e}")
-
-
-def embed_documents(
-    docs: List[Document],
-    embed_model: str = EMBEDDING_MODEL,
-    to_query: bool = False,
-) -> Dict:
-    """
-    将 List[Document] 向量化，转化为 VectorStore.add_embeddings 可以接受的参数
-    """
-    texts = [x.page_content for x in docs]
-    metadatas = [x.metadata for x in docs]
-    embeddings = embed_texts(texts=texts, embed_model=embed_model, to_query=to_query).data
-    if embeddings is not None:
-        return {
-            "texts": texts,
-            "embeddings": embeddings,
-            "metadatas": metadatas,
-        }
